% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date_replicate_range}
\alias{date_replicate_range}
\title{Replicate rows across date range}
\usage{
date_replicate_range(
  data,
  start_date_col,
  end_date_col,
  dest_date_col,
  interval
)
}
\arguments{
\item{data}{Table.}

\item{start_date_col}{Date column of start date.}

\item{end_date_col}{Date column of end date.}

\item{dest_date_col}{Date column to write replicate date.}

\item{interval}{Interval to replicate over. Must be in 'day', 'week', 'month', 'year'.}
}
\value{
Table of replicated rows with all original columns.
}
\description{
Creates a new row for each interval within a date range.
}
\examples{
\dontrun{
 tibble::tribble(
 ~ID , ~start                       , ~end                         ,
   1 , lubridate::ymd('2026-01-01') , lubridate::ymd('2026-01-31') ,
   2 , lubridate::ymd('2026-01-01') , lubridate::ymd('2027-01-01') ,
 ) \%>\%
 replicate_date_range(
   start,
   end,
   dest,
   'day'
 )
}
}
