% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{distance_between_XY}
\alias{distance_between_XY}
\title{Distance between coordinates}
\usage{
distance_between_XY(
  data,
  a.lon_x,
  a.lat_y,
  b.lon_x,
  b.lat_y,
  dist_col = "dist",
  drop_units = TRUE,
  target_crs = 4326
)
}
\arguments{
\item{data}{sf data frame like object.}

\item{a.lon_x}{Column name of first coordinate longitude.}

\item{a.lat_y}{Column name of first coordinate latitude.}

\item{b.lon_x}{Column name of second coordinate longitude.}

\item{b.lat_y}{Column name of second coordinate latitude.}

\item{dist_col}{Column name in which to store calculated distance.}

\item{drop_units}{Column name of second coordinate latitude.}

\item{target_crs}{CRS to use for calculation. Defaults to 4326 (WGS84).}
}
\value{
Input data with \code{dist_col} appended.
}
\description{
Initially from 04a_911_compare.R
}
