% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extract_coords}
\alias{extract_coords}
\title{Extract centroid coordinates to columns}
\usage{
extract_coords(
  data,
  lat_to = "lat",
  lon_to = "lon",
  centroid = FALSE,
  drop_geometry = FALSE
)
}
\arguments{
\item{data}{sf data frame like object.}

\item{lat_to}{Latitude column name, with quotes.}

\item{lon_to}{Longitude column name, with quotes.}

\item{centroid}{Convert the data to centroids first. Defaults to FALSE.}

\item{drop_geometry}{Drop the sf geometry. Defaults to FALSE.}
}
\value{
The input object with new lat and lon columns.
}
\description{
Adds centroid lat and long coordinates as new columns.
}
