% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{geocode_with_geocodeServer}
\alias{geocode_with_geocodeServer}
\title{Geocode with ArcGIS GeocodeServer}
\usage{
geocode_with_geocodeServer(addresses, geocode_url, crs = 4326)
}
\arguments{
\item{addresses}{Vector of single-line address strings to geocode.}

\item{geocode_url}{Full URL of ArcGIS GeocodeServer endpoint.
Should end with \code{GeocodeServer/findAddressCandidates}.}

\item{crs}{CRS geocoder should output. Defaults to 4326 (WGS84).}
}
\value{
Data frame with \code{input_address}, \code{candidate_address}, \code{lat}, \code{lon},
and \code{score} columns
}
\description{
Passess address vector to a ArcGIS GeocodeServer endpoint.
}
\examples{
\dontrun{
# Geocode a vector of addresses with the NC OneMap GeocodeServer
address_vector = c('123 EASY STREET, RALEIGH, NC')
geocode_result <- geocode_with_geocodeServer(
 addresses_vector,
 'https://services.nconemap.gov/secure/rest/services/AddressNC/AddressNC_geocoder/GeocodeServer/findAddressCandidates'
) |> tibble::tibble()
}
}
\keyword{geocode}
