% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_intermediate_rds_store}
\alias{get_intermediate_rds_store}
\title{Get target dir}
\usage{
get_intermediate_rds_store(target, rds_dir)
}
\arguments{
\item{target}{Target file path. Should end in .rds or .csv.}

\item{rds_dir}{The directory to store RDS representations of CSVs.}
}
\value{
returns
}
\description{
Description
}
