% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{is_batch}
\alias{is_batch}
\title{Check if the script is run via R CMD BATCH}
\usage{
is_batch()
}
\value{
Logical
}
\description{
Checks if script is run via R CMD BATCH by looking for --file or -f argument.
}
