% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{open_log}
\alias{open_log}
\title{Open a log file}
\usage{
open_log(name)
}
\arguments{
\item{name}{Name of log file to open or create.}
}
\value{
\code{log} object to pass to \code{write_log()}.
}
\description{
Opens a log file connection and creates the file if it does not exist.
}
\keyword{log}
