% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_csv_from_intermediate}
\alias{read_csv_from_intermediate}
\title{Read a CSV path but from intermediate RDS store.}
\usage{
read_csv_from_intermediate(path, rds_dir = ".csv_overwrites")
}
\arguments{
\item{path}{CSV path to read}

\item{rds_dir}{The directory to store RDS representations of CSVs.}
}
\value{
Returns data read from RDS store.
}
\description{
Description
}
