% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc.R
\name{require_batch}
\alias{require_batch}
\title{Throw error if not run via R CMD BATCH}
\usage{
require_batch()
}
\value{
Logical
}
\description{
Checks if script is run via R CMD BATCH by looking for --file or -f argument.
}
