% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{write_log}
\alias{write_log}
\title{Write to a log file}
\usage{
write_log(message, log_obj, type = "INFO")
}
\arguments{
\item{message}{Message to log.}

\item{log_obj}{\code{log} object created with \code{open_log()}.}

\item{type}{Log message type. Valid values are \code{INFO}, \code{WARN}, and \code{ERROR}.}
}
\value{
Echos log message to console.
}
\description{
Writes a timestamped message to a log file.
}
\keyword{log}
