% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{xy_to_point}
\alias{xy_to_point}
\title{Non-destructive X-Y coordinates to sf point}
\usage{
xy_to_point(data, lon, lat, target_crs = NULL)
}
\arguments{
\item{data}{Tibble that includes lat-long coordinate pairs.}

\item{lon}{Longitude column name, without quotes.}

\item{lat}{Latitude column name, without quotes.}

\item{target_crs}{SF CRS id to set. Default is 4326, which is
returned by many geocding services.}
}
\value{
The input tibble as SF object with geometry column.
}
\description{
Converts tabular coordinates into SF points while preserving the
coordinate columns in the target data.
}
