% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaup-package.R
\docType{package}
\name{yaup-package}
\alias{yaup}
\alias{yaup-package}
\title{yaup: Yet Another Utility Package}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

YAUP is a catch-all package for utility functions that I use across different projects. This package is primarily for my own personal use, and I make no claims or guarantees about its quality or maintenance.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://jpcollins.me/yaup/}
}

}
\author{
\strong{Maintainer}: James Collins \email{jpco@unc.edu} (\href{https://orcid.org/0000-0001-5751-9733}{ORCID})

}
\keyword{internal}
