% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{confirm_overwrite}
\alias{confirm_overwrite}
\title{Confirm before overwriting existing output file with changed data}
\usage{
confirm_overwrite(
  data,
  target,
  force_in_batch = TRUE,
  compress = "none",
  pass_data = FALSE,
  archive = NULL,
  rds_dir = ".csv_overwrites",
  use_here = TRUE
)
}
\arguments{
\item{data}{Current working data.}

\item{target}{Target file path. Should end in .rds or .csv.}

\item{force_in_batch}{If called via an R CMD BATCH job, whether to force overwrite and archive. Default TRUE.}

\item{compress}{Pass-through parameter to readr::write_rds.}

\item{pass_data}{If successful, return the supplied data for further action.
Default true.}

\item{archive}{If user chooses to overwrite file, rename existing file as
archival copy. Default NULL prompts for action.}

\item{rds_dir}{The directory to store RDS representations of CSVs.}

\item{use_here}{Use here::here() to reference to project.}
}
\value{
returns
}
\description{
Description
}
\examples{
\dontrun{
# Example title
}
}
