% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{geocode_with_geocodeServer_multiCandidate}
\alias{geocode_with_geocodeServer_multiCandidate}
\title{Geocode with ArcGIS GeocodeServer and return multiple candidate}
\usage{
geocode_with_geocodeServer_multiCandidate(
  data,
  address_col,
  city_col,
  state_col,
  geocode_url,
  id_col = "row_id",
  crs = 4326,
  prefix = "gc."
)
}
\arguments{
\item{data}{Data frame-like object containing columns with addresses, city
names, and stae names.}

\item{address_col}{Column name containing street addresses.}

\item{city_col}{Column name containing city names.}

\item{state_col}{Column name containing state names.}

\item{geocode_url}{Full URL of ArcGIS GeocodeServer endpoint. Should end
with \code{GeocodeServer/findAddressCandidates}.}

\item{id_col}{Column name containing unique identifiers. Defaults to
\code{row_id}.}

\item{crs}{CRS geocoder should output. Defaults to 4326 (WGS84).}

\item{prefix}{Prefix for geocoder result columns. Defaults to \code{gc.}.}
}
\value{
Tibble with geocode results.
}
\description{
Passess address vector to a ArcGIS GeocodeServer endpoint. Returns multiple
candidates for an input row when available.
}
\examples{
\dontrun{
# Geocode a vector of addresses with the NC OneMap GeocodeServer
addresses = tibble::tibble(
 row_id = c(1),
 address = c('123 EASY STREET'),
 city = c('RALEIGH'),
 state = c('NC')
)

geocode_result <- addresses |>
 geocode_with_geocodeServer(
   address,
   city,
   state,
   'https://services.nconemap.gov/secure/rest/services/AddressNC/AddressNC_geocoder/GeocodeServer/findAddressCandidates'
 )
}
}
\keyword{geocode}
