% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{write_log}
\alias{write_log}
\title{Write to a log file}
\usage{
write_log(message, log_obj, type = "INFO")
}
\arguments{
\item{name}{description}
}
\value{
returns
}
\description{
Description
}
\examples{
\dontrun{
# Example title
}
}
